requirements.txt
torch>=2
torchvision
pyyaml
numpy
opencv-python
submitit
braceexpand
webdataset
timm
decord
pandas
einops
beartype
wandb
scikit-image
tqdm
mesh2sdf
open3d